# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Author {
  name: string;
  /** @nullable */
  affiliation: string | null;
  /** @nullable */
  profileUrl: string | null;
  sortOrder: number;
}

export interface ExternalLink {
  label: string;
  url: string;
}

export interface PaperSummary {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  shortTitle: string | null;
  /**
     * Short description used on research cards and as a search snippet
     * @maxLength 20000
     * @nullable
     */
  summary: string | null;
  /**
     * Full abstract shown on the paper page
     * @maxLength 20000
     */
  abstract: string;
  status: string;
  /** @nullable */
  venue: string | null;
  authors: Author[];
  /** @nullable */
  submittedDate: string | null;
  year: number;
  keywords: string[];
  featured: boolean;
  published: boolean;
  updatedAt: string;
}

export type Paper = PaperSummary & ({
  /** Sanitized rich-text HTML */
  content: string;
  externalLinks: ExternalLink[];
  /** @nullable */
  seoImage: string | null;
  createdAt: string;
});

export interface AuthorInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  affiliation: string | null;
  /** @nullable */
  profileUrl: string | null;
}

export interface ExternalLinkInput {
  /** @minLength 1 */
  label: string;
  url: string;
}

export interface PaperInput {
  /**
     * @minLength 1
     * @maxLength 180
     */
  slug: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  title: string;
  /** @nullable */
  shortTitle: string | null;
  /**
     * @maxLength 20000
     * @nullable
     */
  summary: string | null;
  /** @maxLength 20000 */
  abstract: string;
  content: string;
  status: string;
  /** @nullable */
  venue: string | null;
  authors: AuthorInput[];
  /** @nullable */
  submittedDate: string | null;
  year: number;
  keywords: string[];
  featured: boolean;
  published: boolean;
  externalLinks: ExternalLinkInput[];
  /** @nullable */
  seoImage: string | null;
}

export type PaperUpdate = PaperInput;

export interface LoginInput {
  /** @minLength 1 */
  password: string;
}

export interface SessionState {
  authenticated: boolean;
}

export interface DashboardSummary {
  totalPapers: number;
  publishedPapers: number;
  drafts: number;
  recentlyUpdated: PaperSummary[];
}

export interface PublishInput {
  published: boolean;
}

export interface ImageUploadInput {
  image: Blob;
}

export interface ImageUploadResult {
  url: string;
  publicId: string;
  width: number;
  height: number;
}

export interface Success {
  success: boolean;
}