# lib/api-zod/src/generated/types/authorInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthorInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  affiliation: string | null;
  /** @nullable */
  profileUrl: string | null;
}