# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaperSummary } from './paperSummary';

export interface DashboardSummary {
  totalPapers: number;
  publishedPapers: number;
  drafts: number;
  recentlyUpdated: PaperSummary[];
}