# lib/api-zod/src/generated/types/externalLinkInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExternalLinkInput {
  /** @minLength 1 */
  label: string;
  url: string;
}