# lib/api-zod/src/generated/types/imageUploadResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ImageUploadResult {
  url: string;
  publicId: string;
  width: number;
  height: number;
}