# lib/api-zod/src/generated/types/paper.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExternalLink } from './externalLink';
import type { PaperSummary } from './paperSummary';

export type Paper = PaperSummary & ({
  /** Sanitized rich-text HTML */
  content: string;
  externalLinks: ExternalLink[];
  /** @nullable */
  seoImage: string | null;
  createdAt: Date;
});