# lib/api-zod/src/generated/types/paperUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PaperInput } from './paperInput';

export type PaperUpdate = PaperInput;